Drop support for SPDY/3.
It's redundant with HTTP/2.
