+++ okcurl/README.md
@@ -4,4 +4,4 @@ OkCurl
 _A curl for the next-generation web._
 
 OkCurl is an OkHttp-backed curl clone which allows you to test OkHttp's HTTP engine (including
-SPDY and HTTP/2) against web servers.
+HTTP/2) against web servers.
